Allow overriding the graph during tests
All of the graph state and complexity has been moved into the graph
state object, leaving the graph object as a simple wrapper for the nicer
user interfaces. This also allows the state of the graph to be swapped
out without replacing the instance, which is how overrides are being
achieved. The documentation has been updated to relfect this.